casoEstudioFinal/requirements.txt
# Simulador de Cifrados Clásicos vs Modernos
# Dependencias del proyecto

# Criptografía
pycryptodome==3.23.0

# Análisis de datos
pandas==2.3.0
numpy==2.3.0

# Visualización
matplotlib==3.10.3
seaborn==0.13.2

# Utilidades adicionales
scipy==1.15.3